simplify and address comments
