+++ tests/unit/source_code/linters/test_spark_connect.py
@@ -1,7 +1,7 @@
 from itertools import chain
 
 from databricks.labs.ucx.source_code.base import Failure
-from databricks.labs.ucx.source_code.linters.python_ast import Tree, TreeWalker
+from databricks.labs.ucx.source_code.linters.python_ast import Tree
 from databricks.labs.ucx.source_code.linters.spark_connect import LoggingMatcher, SparkConnectLinter
 
 
@@ -222,7 +222,7 @@ sc._jvm.org.apache.log4j.LogManager.getLogger(__name__).info("test")
             end_line=7,
             end_col=24,
         ),
-    ] == list(chain.from_iterable([logging_matcher.lint(node) for node in TreeWalker.walk(Tree.parse(code).root)]))
+    ] == list(chain.from_iterable([logging_matcher.lint(node) for node in Tree.parse(code).walk()]))
 
 
 def test_logging_serverless():
@@ -251,7 +251,7 @@ log4jLogger = sc._jvm.org.apache.log4j
             end_line=3,
             end_col=38,
         ),
-    ] == list(chain.from_iterable([logging_matcher.lint(node) for node in TreeWalker.walk(Tree.parse(code).root)]))
+    ] == list(chain.from_iterable([logging_matcher.lint(node) for node in Tree.parse(code).walk()]))
 
 
 def test_valid_code():
